requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python-headless==4.10.0.84
pillow==10.3.0
Werkzeug==3.0.3
zipp==3.19.2
